feat: add utils and move some files to interfaces
