+++ projects/angular/src/forms/datepicker/date-container.spec.ts
@@ -20,10 +20,10 @@ import { LocaleHelperService } from './providers/locale-helper.service';
 import { ViewManagerService } from './providers/view-manager.service';
 import { TestContext } from '../../data/datagrid/helpers.spec';
 import { ClrPopoverPositions } from '../../popover/common/enums/positions.enum';
-import { PopoverPosition } from '../../popover/common/popover-positions';
 import { ClrPopoverEventsService } from '../../popover/common/providers/popover-events.service';
 import { ClrPopoverPositionService } from '../../popover/common/providers/popover-position.service';
 import { ClrPopoverToggleService } from '../../popover/common/providers/popover-toggle.service';
+import { PopoverPosition } from '../../popover/common/utils/popover-positions';
 import { expectActiveElementToBe } from '../../utils/testing/helpers.spec';
 import { ClrCommonFormsModule } from '../common/common.module';
 import { CONTROL_STATE, IfControlStateService } from '../common/if-control-state/if-control-state.service';
