+++ projects/angular/src/popover/common/abstract-popover.ts
@@ -19,8 +19,8 @@ import {
 import { Subscription } from 'rxjs';
 import { startWith } from 'rxjs/operators';
 
+import { PopoverOptions } from './interfaces/popover-options.interface';
 import { Point, Popover } from './popover';
-import { PopoverOptions } from './popover-options.interface';
 import { ClrPopoverToggleService } from '../../popover/common/providers/popover-toggle.service';
 import { Keys } from '../../utils/enums/keys.enum';
 import { normalizeKey } from '../../utils/focus/key-focus/util';
